another debug for beast

+++ eval_with_beast2.py
@@ -189,9 +189,9 @@ with open(output_file, mode='w', newline='', encoding='utf-8') as f:
                 row["char_count"]
             ])
             lab = int(row["class_label"])
-            wer = int(row["wer"])
+            wer = float(row["wer"])
             word_N = int(row["word_count"])
-            cer = int(row["cer"])
+            cer = float(row["cer"])
             char_N = int(row["char_count"])
             all_wer_per_class[lab].extend([wer])  # Add the current batch's WERs to the list
             all_wN_per_class[lab].extend([word_N])  # Add the current batch's WERs to the list
